`GetUserCommand()` moved to `TextEditor` new naming
naming of `IsInputSizeValid()` changed to `IfInputSizeNotValidClearSTDin()`
